change aspnet testapplication name to match the pattern
