+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/.dockerignore
