+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/Directory.Build.props
