+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/Global.asax
@@ -1,0 +1,1 @@
+﻿<%@ Application Codebehind="Global.asax.cs" Inherits="TestApplication.AspNet.NetFramework.MvcApplication" Language="C#" %>
