+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/Helpers/AssembliesHelper.cs
@@ -18,7 +18,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace TestApplication.AspNet.Helpers;
+namespace TestApplication.AspNet.NetFramework.Helpers;
 
 public static class AssembliesHelper
 {
