+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/Properties/AssemblyInfo.cs
@@ -21,11 +21,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("TestApplication.AspNet")]
+[assembly: AssemblyTitle("TestApplication.AspNet.NetFramework")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("TestApplication.AspNet")]
+[assembly: AssemblyProduct("TestApplication.AspNet.NetFramework")]
 [assembly: AssemblyCopyright("Copyright ©  2021")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
