+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/Properties/PublishProfiles/FolderProfile.Debug.pubxml
