+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/Views/Home/Index.cshtml
