+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/Views/Shared/_Layout.cshtml
@@ -1,14 +1,14 @@
 ﻿<!DOCTYPE html>
 <html>
 <head>
-    <title>TestApplication.AspNet</title>
+    <title>TestApplication.AspNet.NetFramework</title>
 </head>
 <body>
     <div class="body-content">
         @RenderBody()
         <hr />
         <footer>
-            <p>&copy; @DateTime.Now.Year - TestApplication.AspNet</p>
+            <p>&copy; @DateTime.Now.Year - TestApplication.AspNet.NetFramework</p>
         </footer>
     </div>
 </body>
