+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/Views/Web.config
@@ -16,7 +16,7 @@
         <add namespace="System.Web.Mvc.Ajax" />
         <add namespace="System.Web.Mvc.Html" />
         <add namespace="System.Web.Routing" />
-        <add namespace="TestApplication.AspNet" />
+        <add namespace="TestApplication.AspNet.NetFramework" />
       </namespaces>
     </pages>
   </system.web.webPages.razor>
