+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/Views/_ViewStart.cshtml
