+++ test/test-applications/integrations/TestApplication.AspNet.NetFramework/packages.config
