--- test/test-applications/integrations/TestApplication.AspNet/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="TestApplication.AspNet.MvcApplication" Language="C#" %>
